修复最后一个下载线程中对Rang范围计算错误导致非均匀分配下载量时会下载失败的BUG(fileSize%threadNum除不尽时)

+++ library/src/main/java/com/lijunhuayc/downloader/downloader/DownloadProgressListener.java
@@ -6,7 +6,7 @@ package com.lijunhuayc.downloader.downloader;
  * Email: lijunhuayc@sina.com
  */
 public interface DownloadProgressListener {
-    void onDownloadTotalSize(int totalSize);
+    void onDownloadTotalSize(long totalSize);
 
     /**
      * Real-time update downloading progress
@@ -15,7 +15,7 @@ public interface DownloadProgressListener {
      * @param percent downloading percent(%)
      * @param speed   downloading speed(KB/S)
      */
-    void updateDownloadProgress(int size, float percent, float speed);
+    void updateDownloadProgress(long size, float percent, float speed);
 
     void onDownloadSuccess(String apkPath);
